Fix checkin toggle animation
Before the check-in state was stored in the db and
`isLoading` as a react state. This would sometimes
cause react to see new check-in state while `isLoading`
was still true which made the check-in animation choppy.

This stores the loading state together with the check-in
state inside the DB so that states are always in sync with
each other.

We just need to make sure to reset the loading state
when the page is first loaded in case the tab/browser
had been closed unexpectedly while it was loading.